requirements.txt
flask>=3.0.0
twilio>=9.0.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.30.0
google-generativeai>=0.7.0
